+++ src/array/reducer/set-user-defined-property.spec.ts
@@ -1,6 +1,6 @@
 import { SetUserDefinedPropertyAction } from '../../actions';
 import { setUserDefinedPropertyReducer } from './set-user-defined-property';
-import { FORM_CONTROL_ID, INITIAL_STATE } from './test-util';
+import { FORM_CONTROL_ID, FORM_CONTROL_0_ID, INITIAL_STATE } from './test-util';
 
 describe(`form group ${setUserDefinedPropertyReducer.name}`, () => {
   it('should skip any actionof the wrong type', () =>
@@ -21,5 +21,14 @@ describe(`form group ${setUserDefinedPropertyReducer.name}`, () => {
     const state = { ...INITIAL_STATE, userDefinedProperties: { [prop]: value } };
     const resultState = setUserDefinedPropertyReducer(state, new SetUserDefinedPropertyAction(FORM_CONTROL_ID, prop, value));
     expect(resultState).toBe(state);
+  });
+
+  it('should update state user defined properties for children', () => {
+    const prop = 'prop';
+    const value = 12;
+    const resultState = setUserDefinedPropertyReducer(INITIAL_STATE, new SetUserDefinedPropertyAction(FORM_CONTROL_0_ID, prop, value));
+    expect(resultState.controls[0].userDefinedProperties).toEqual({
+      [prop]: value,
+    });
   });
 });