+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## ngrx-forms Changelog
 
+<a name="2.1.1"></a>
+### 2.1.1
+
+#### Bugfixes
+
+* fix `select` controls not properly selecting `option` if `option` is added to the DOM after the value of the state was set to the value of the option, closes [#23](https://github.com/MrWolfZ/ngrx-forms/issues/23)
+
 <a name="2.1.0"></a>
 ### 2.1.0